fix(ntn): satisfy CI's flake8 and the test-inclusive typecheck
E129 on a wrapped return annotation, and a ByteSource|null decode in the new
api test. Both are CI-only: the repo's flake8 is stricter than the local hook,
and `pnpm -r typecheck` covers test files that a bare -p tsconfig run skips.

+++ typescript/packages/core/src/commands/cli/builtin/ntn/ntn.test.ts
@@ -259,7 +259,7 @@ describe('ntn verbs', () => {
     // the engine's own parse message nor a generic usage error.
     const [, io] = unwrap(await api(makeInv({ data: '{' }, ['v1/search'])))
     expect(io.exitCode).toBe(1)
-    expect(DEC.decode(io.stderr)).toBe('error: Invalid JSON from --data\n')
+    expect(DEC.decode(io.stderr as Uint8Array)).toBe('error: Invalid JSON from --data\n')
   })
 
   it('api builds nested bodies from bracket paths and typed assignments', async () => {
